# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "立言阁 Server API",
    "description": "The server-owned API contract for 立言阁.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/task-creation/prepare": {
      "post": {
        "tags": [
          "task creation"
        ],
        "summary": "Prepare Task Source",
        "operationId": "prepare_task_source",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrepareSourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/task-creation/confirm": {
      "post": {
        "tags": [
          "task creation"
        ],
        "summary": "Confirm Task Creation",
        "operationId": "confirm_task_creation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmTaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfirmTaskRequest": {
        "properties": {
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "client_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Session Id"
          },
          "source_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Source Ids"
          },
          "accepted_warning_versions": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "format": "uuid"
            },
            "type": "object",
            "title": "Accepted Warning Versions"
          }
        },
        "type": "object",
        "required": [
          "idempotency_key"
        ],
        "title": "ConfirmTaskRequest"
      },
      "ConfirmTaskResponse": {
        "properties": {
          "task": {
            "$ref": "#/components/schemas/TaskSummary"
          },
          "source_revision": {
            "$ref": "#/components/schemas/SourceRevisionResponse"
          },
          "source_revisions": {
            "items": {
              "$ref": "#/components/schemas/SourceRevisionResponse"
            },
            "type": "array",
            "title": "Source Revisions"
          }
        },
        "type": "object",
        "required": [
          "task",
          "source_revision",
          "source_revisions"
        ],
        "title": "ConfirmTaskResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PreparationWarning": {
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "short_body",
              "missing_provenance"
            ],
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "PreparationWarning"
      },
      "PrepareSourceResponse": {
        "properties": {
          "source": {
            "$ref": "#/components/schemas/PreparedSource"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/PreparationWarning"
            },
            "type": "array",
            "title": "Warnings"
          },
          "can_confirm": {
            "type": "boolean",
            "const": true,
            "title": "Can Confirm",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "source",
          "warnings"
        ],
        "title": "PrepareSourceResponse"
      },
      "PreparedSource": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "provenance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provenance"
          }
        },
        "type": "object",
        "required": [
          "title",
          "body",
          "provenance"
        ],
        "title": "PreparedSource"
      },
...
      "SourceInput": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "provenance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provenance"
          }
        },
        "type": "object",
        "required": [
          "title",
          "body"
        ],
        "title": "SourceInput"
      },
...
      "SourceRevisionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "provenance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provenance"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "body",
          "provenance"
        ],
        "title": "SourceRevisionResponse"
      },
...
      "TaskSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "first_source_title": {
            "type": "string",
            "title": "First Source Title"
          },
          "additional_source_count": {
            "type": "integer",
            "title": "Additional Source Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_activity_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Activity At"
          },
          "current_version_id": {
            "type": "string",
            "title": "Current Version Id"
          },
          "current_version_number": {
            "type": "integer",
            "title": "Current Version Number"
          },
          "can_delete": {
            "type": "boolean",
            "title": "Can Delete"
          },
          "delete_disabled_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delete Disabled Reason"
          }
        },
        "type": "object",
        "required": [
          "id",
          "number",
          "display_name",
          "first_source_title",
          "additional_source_count",
          "created_at",
          "last_activity_at",
          "current_version_id",
          "current_version_number",
          "can_delete",
          "delete_disabled_reason"
        ],
        "title": "TaskSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}